비용 관련 필드 Long > BigDecimal 변경

+++ src/main/java/com/devtaste/facampay/application/payment/dto/PaymentDTO.java
@@ -3,15 +3,16 @@ package com.devtaste.facampay.application.payment.dto;
 import com.devtaste.facampay.domain.model.payment.Payment;
 import com.devtaste.facampay.domain.model.payment.type.PaymentStatusType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record PaymentDTO(Long paymentId, Long money, PaymentStatusType paymentStatus, LocalDateTime createdAt) {
+public record PaymentDTO(Long paymentId, BigDecimal money, PaymentStatusType paymentStatus, LocalDateTime createdAt) {
 
     public static PaymentDTO from(Payment payment) {
         return new PaymentDTO(payment.getPaymentId(), payment.getMoney(), payment.getPaymentStatus(), payment.getCreatedAt());
     }
 
-    public static PaymentDTO of(Long paymentId, Long money, PaymentStatusType paymentStatus, LocalDateTime createdAt) {
+    public static PaymentDTO of(Long paymentId, BigDecimal money, PaymentStatusType paymentStatus, LocalDateTime createdAt) {
         return new PaymentDTO(paymentId, money, paymentStatus, createdAt);
     }
 }
